feat: presents ranked ballots like we do to closed option ballots
If we have m candidates, each completed ballot contains an array with m elements.
If the voter gives a formal ballot, the array contains the identifiers of the selected candidates in the ranked order.
If the voter gives a blank ballot, the array contains in each position the value m.
If the voter gives a null ballot, the array contains in each position the value m+1.

Important! This change is followed by a change in the operative backend.