feat: upgrades the votes by time statistics
Adds the accumulated votes in the time chart
Generalizes the classic selector component, now it receives the data that must been showed, but doesn't modify the data.
The data must been parsed on the father component depending on the type of data.
The charts shows the dateTime labels depending on the selected deltaTime
Updates the section design.