feat: shows different statistics depending on the poll and the groups
When there is no poll, we can only know about the participation and the time distribution.
When there is poll and there are groups we can know about all the statistics.
When there is poll and there are groups we can resume voters and poll statistics into weights.
This commit replaces the hardcoded values too.